Don't fetch member list if not affiliated
Fixes #1426
