+++ src/headless/plugins/muc/constants.js
@@ -1,0 +1,6 @@
+export const MUC_ROLE_WEIGHTS = {
+    'moderator': 1,
+    'participant': 2,
+    'visitor': 3,
+    'none': 2
+};
