+++ src/headless/plugins/muc/occupants.js
@@ -1,16 +1,10 @@
 import ChatRoomOccupant from './occupant.js';
 import u from '../../utils/form';
 import { Collection } from '@converse/skeletor/src/collection';
+import { MUC_ROLE_WEIGHTS } from './constants.js';
 import { Strophe } from 'strophe.js/src/strophe';
 import { _converse, api } from '../../core.js';
 import { getAffiliationList } from './affiliations/utils.js';
-
-const MUC_ROLE_WEIGHTS = {
-    'moderator': 1,
-    'participant': 2,
-    'visitor': 3,
-    'none': 2
-};
 
 
 /**
@@ -34,12 +28,26 @@ const ChatRoomOccupants = Collection.extend({
         }
     },
 
+    /**
+     * Get the {@link _converse.ChatRoomOccupant} instance which
+     * represents the current user.
+     * @method _converse.ChatRoomOccupants#getOwnOccupant
+     * @returns { _converse.ChatRoomOccupant }
+     */
+    getOwnOccupant () {
+        return this.findWhere({ 'jid': _converse.bare_jid });
+    },
+
     getAutoFetchedAffiliationLists () {
         const affs = api.settings.get('muc_fetch_members');
         return Array.isArray(affs) ? affs : affs ? ['member', 'admin', 'owner'] : [];
     },
 
     async fetchMembers () {
+        if (!['member', 'admin', 'owner'].includes(this.getOwnOccupant()?.get('affiliation'))) {
+            // https://xmpp.org/extensions/xep-0045.html#affil-priv
+            return;
+        }
         const affiliations = this.getAutoFetchedAffiliationLists();
         if (affiliations.length === 0) {
             return;
@@ -59,26 +67,18 @@ const ChatRoomOccupants = Collection.extend({
                 !new_jids.includes(m.get('jid'))
             );
         });
-
         removed_members.forEach(occupant => {
             if (occupant.get('jid') === _converse.bare_jid) {
                 return;
-            }
-            if (occupant.get('show') === 'offline') {
+            } else if (occupant.get('show') === 'offline') {
                 occupant.destroy();
             } else {
                 occupant.save('affiliation', null);
             }
         });
         new_members.forEach(attrs => {
-            const occupant = attrs.jid
-                ? this.findOccupant({ 'jid': attrs.jid })
-                : this.findOccupant({ 'nick': attrs.nick });
-            if (occupant) {
-                occupant.save(attrs);
-            } else {
-                this.create(attrs);
-            }
+            const occupant = this.findOccupant(attrs);
+            occupant ? occupant.save(attrs) : this.create(attrs);
         });
         /**
          * Triggered once the member lists for this MUC have been fetched and processed.
@@ -101,7 +101,6 @@ const ChatRoomOccupants = Collection.extend({
      * If we have a JID, we use that as lookup variable,
      * otherwise we use the nick. We don't always have both,
      * but should have at least one or the other.
-     * @private
      * @method _converse.ChatRoomOccupants#findOccupant
      * @param { OccupantData } data
      */
